+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- Remove `trimPyramid` and require declaration of `tileSize` in `ZarrPixelSource`.
+
 ## 0.9.1
 
 ### Added